requirements.txt
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==3.25.0
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
-e git+https://github.com/buzzromain/projet-web-serveur.git@941df0b7ab922544d4e264df32a51b50913dabd1#egg=projet_web_serveur
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
